+++ trace/src/reflection/SerializeTypes.h
@@ -624,7 +624,7 @@ namespace trace::Reflection {
 	void SerializeContainer(std::map<Key, T, Compare, Alloc>& object, void* location, void* member_info, uint16_t format)
 	{
 		using key_type = std::remove_cv_t<Key>;
-		uint32_t size = object.size();
+		uint32_t size = static_cast<uint32_t>(object.size());
 		constexpr std::string_view type_name = TypeName< std::map<Key, T, Compare, Alloc>>();
 		BeginKeyValueContainer(type_name,location,member_info, format);
 		SerializeContainerSize(size, location, member_info, format);
